구독 시스템 구축 (#20)
* feat: 간단한 구독 반영

* refactor: amount int를 long으로 변경

* refactor: 빠진 부분 수정

* refactor: orderId null 이슈 미리 수정

* feat: 토스 연동 (구독은 아님)

* fix: 토스도 구독 되게 변경

* refactor: 반환값 작성

* feat: 이제 토스도 키값을 이용해서 구독 서비스 이용

+++ apps/dragons_api/src/main/java/com/dragons/application/payment/dto/PaymentBankTransferResult.java
@@ -4,7 +4,7 @@ public record PaymentBankTransferResult(
     String bankCode,
     String accountNumber,
     String depositorName,
-    int amount,
+    long amount,
     String planType
 ) {
